google_sheets_to_mysql/requirements.txt
boto3==1.34.119
google_api_python_client==2.128.0
gspread==6.1.4
mysql-connector-python==9.1.0
pymysql==1.1.1
mysql_connector_repackaged==0.3.1
numpy==2.2.1
pandas==2.2.3
protobuf==5.29.3
SQLAlchemy==2.0.30